+++ bmc-database/src/main/java/com/oracle/bmc/database/DatabaseAsync.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
  */
 package com.oracle.bmc.database;
 
@@ -365,7 +365,7 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Creates a new database home in the specified DB system based on the request parameters you provide.
+     * Creates a new Database Home in the specified DB system based on the request parameters you provide. Applies only to bare metal and Exadata DB systems.
      *
      *
      * @param request The request object containing the details to send
@@ -545,7 +545,7 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Deletes a DB Home. The DB Home and its database data are local to the DB system and will be lost when it is deleted. Oracle recommends that you back up any data in the DB system prior to deleting it.
+     * Deletes a Database Home. The Database Home and its database data are local to the DB system and are lost when you delete the Database Home. Oracle recommends that you back up any data on the DB system before you delete it.
      *
      *
      * @param request The request object containing the details to send
@@ -950,7 +950,7 @@ public interface DatabaseAsync extends AutoCloseable {
             com.oracle.bmc.responses.AsyncHandler<GetDatabaseRequest, GetDatabaseResponse> handler);
 
     /**
-     * Gets information about the specified database home.
+     * Gets information about the specified Database Home.
      *
      * @param request The request object containing the details to send
      * @param handler The request handler to invoke upon completion, may be null.
@@ -1414,7 +1414,7 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Gets a list of the databases in the specified database home.
+     * Gets a list of the databases in the specified Database Home.
      *
      *
      * @param request The request object containing the details to send
@@ -1430,7 +1430,7 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Gets history of the actions taken for patches for the specified database home.
+     * Gets history of the actions taken for patches for the specified Database Home.
      *
      *
      * @param request The request object containing the details to send
@@ -1449,7 +1449,7 @@ public interface DatabaseAsync extends AutoCloseable {
                             handler);
 
     /**
-     * Lists patches applicable to the requested database home.
+     * Lists patches applicable to the requested Database Home.
      *
      *
      * @param request The request object containing the details to send
@@ -1466,7 +1466,7 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Gets a list of database homes in the specified DB system and compartment. A database home is a directory where Oracle Database software is installed.
+     * Gets a list of Database Homes in the specified DB system and compartment. A Database Home is a directory where Oracle Database software is installed.
      *
      *
      * @param request The request object containing the details to send