+++ bmc-database/src/main/java/com/oracle/bmc/database/model/AutonomousContainerDatabaseSummary.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
  */
 package com.oracle.bmc.database.model;
 
@@ -125,6 +125,15 @@ public class AutonomousContainerDatabaseSummary {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("maintenanceWindow")
+        private MaintenanceWindow maintenanceWindow;
+
+        public Builder maintenanceWindow(MaintenanceWindow maintenanceWindow) {
+            this.maintenanceWindow = maintenanceWindow;
+            this.__explicitlySet__.add("maintenanceWindow");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
         private java.util.Map<String, String> freeformTags;
 
@@ -179,6 +188,7 @@ public class AutonomousContainerDatabaseSummary {
                             patchModel,
                             lastMaintenanceRunId,
                             nextMaintenanceRunId,
+                            maintenanceWindow,
                             freeformTags,
                             definedTags,
                             availabilityDomain,
@@ -202,6 +212,7 @@ public class AutonomousContainerDatabaseSummary {
                             .patchModel(o.getPatchModel())
                             .lastMaintenanceRunId(o.getLastMaintenanceRunId())
                             .nextMaintenanceRunId(o.getNextMaintenanceRunId())
+                            .maintenanceWindow(o.getMaintenanceWindow())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags())
                             .availabilityDomain(o.getAvailabilityDomain())
@@ -428,6 +439,9 @@ public class AutonomousContainerDatabaseSummary {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("nextMaintenanceRunId")
     String nextMaintenanceRunId;
+
+    @com.fasterxml.jackson.annotation.JsonProperty("maintenanceWindow")
+    MaintenanceWindow maintenanceWindow;
 
     /**
      * Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace.