+++ bmc-database/src/main/java/com/oracle/bmc/database/model/DataGuardAssociationSummary.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
  */
 package com.oracle.bmc.database.model;
 
@@ -358,7 +358,7 @@ public class DataGuardAssociationSummary {
     String peerDbSystemId;
 
     /**
-     * The [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the database home containing the associated peer database.
+     * The [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the Database Home containing the associated peer database.
      *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("peerDbHomeId")