+++ bmc-database/src/main/java/com/oracle/bmc/database/model/UpdateDbHomeDetails.java
@@ -1,10 +1,10 @@
 /**
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
  */
 package com.oracle.bmc.database.model;
 
 /**
- * Describes the modification parameters for the database home.
+ * Describes the modification parameters for the Database Home.
  *
  * <br/>
  * Note: Objects should always be created or deserialized using the {@link Builder}. This model distinguishes fields