+++ bmc-examples/src/main/java/GetInstancePublicIpExample.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
  */
 import java.util.ArrayList;
 import java.util.HashSet;